+++ ivy/functional/backends/paddle/experimental/layers.py
@@ -1,7 +1,7 @@
 # global
+from turtle import st
 from typing import Optional, Union, Tuple, List, Literal, Sequence, Callable
 import paddle
-from hypothesis import given
 
 from ivy.functional.ivy.layers import (
     _handle_padding,
@@ -665,6 +665,7 @@ def sliding_window(
     )
 
 
+@st.composite
 def interpolate_linear(
     x: paddle.Tensor,
     size: Union[Sequence[int], int],
@@ -679,19 +680,3 @@ def interpolate_linear(
     return paddle.nn.functional.interpolate(
         x, size, scale_factor, mode, align_corners, align_mode, data_format, name
     )
-
-
-@given()
-def test_interpolate_linear(
-    x: paddle.Tensor,
-    size: Union[Sequence[int], int],
-    mode: Optional[Literal["linear", "bilinear", "trilinear"]] = "linear",
-    scale_factor: Optional[Union[Sequence[int], int]] = None,
-    align_corners: Optional[bool] = False,
-    align_mode: int = 0,
-    data_format: str = "NCHW",
-    name: Optional[str] = None,
-):
-    assert paddle.nn.functional.interpolate(
-        x, size, scale_factor, mode, align_corners, align_mode, data_format, name
-    ) == interpolate_linear(x)
